mac os error fixed

+++ src/libblockchain/node.cpp
@@ -1086,7 +1086,7 @@ void block_worker(detail::node_internals& impl)
             }
 
             if (approve > reject &&
-                poll_participants >= std::max(2ull, impl.m_p2p_peers.size() / 2 ))
+                poll_participants >= std::max(2ull, static_cast<uint64_t>(impl.m_p2p_peers.size() / 2 )))
             {
                 if (impl.all_sync_info.headers_actions_data.end() != it_scan_least_revert_approve_winner &&
                     it_scan_least_revert_approve_winner->second.headers.front() != it->second.headers.front())
@@ -1099,7 +1099,7 @@ void block_worker(detail::node_internals& impl)
 
             if (approve > scan_most_approved_revert &&
                 approve > own_vote.first &&
-                poll_participants >= std::max(2ull, impl.m_p2p_peers.size() / 2) &&
+                poll_participants >= std::max(2ull, static_cast<uint64_t>(impl.m_p2p_peers.size() / 2 )) &&
                 (
                     impl.all_sync_info.headers_actions_data.end() == it_scan_most_approved_revert ||
                     it_scan_most_approved_revert->second.reverts_required > revert_coefficient
